Move KeyboardVisibilityUtil to common
